+++ commons/src/main/scala/commons.scala
@@ -63,7 +63,7 @@ object types {
   private val domainFrom = ".+@(.+)"       .r
   private val userFrom   = "(.+)@.+"       .r
   private val domain     = "[\\w\\.\\-]+"  .r
-  private val email      = "[\\w\\.\\-@]+" .r
+  private val email      = """^[A-Za-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[A-Za-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[A-Za-z0-9](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?\.)+[A-Za-z]{2,}$""".r
 
   given JsonEncoder      [TenantName]      = JsonEncoder.string
   given JsonEncoder      [TenantCode]      = JsonEncoder.string